fix installer: wrong detection of hot signer usage.
We did not check that the keys in the primary path
were using the hot signer. In the case of a hot signing
key only used for the primary path,
the hot signer was not advertised to the context
passing to each installer step and was not stored after
install.