Revert "Merge #742: commands: don't add derivation paths for keys from different path but same signer"
This reverts commit b154bcaabf343908f602b5a88668f97dcf82b236, reversing
changes made to de887218b43e0a6e6e2b88454fb57d4ed7d2a823.

This reverts the fixes to the pruning of BIP32 derivation paths when
creating a PSBT, in preparation of reverting the merge of this feature
altogether. This is because always creating PSBT with only the BIP32
derivations for a single path broke the Ledger support.