Revert "Merge #722: Only include BIP32 derivations for a single spending path when creating PSBTs"
This reverts commit 1dc2badd2ffbf191576672bd12e4507039f7a834, reversing
changes made to 86f5e1e3eac0e81bc9e6ecbf49037d9a558bfd90.

This reverts PR #722. It turns out the Ledger Bitcoin app needs the
BIP32 derivation for all the keys in the Script, not only for the
spending path used. Therefore always create PSBT with all the BIP32
derivations. We'll add a way to prune them for talking to the Bitbox in
a future commit.