commands: add `create_spend_internal` function
This is a more general version of `create_spend` in that it allows
for a mixure of both mandatory and optional coin selection candidates
and has a `min_fee` parameter.

Another difference is that it takes destination addresses that have
been checked for the network and amounts as `Amount`.

`create_spend` has been updated to call `create_spend_internal`.