Remove legacy harware_wallets field from gui config
Ledger hmacs, wallet name and fingerprint labels where
moved one year ago in the settings.json file.
Here the commit of the new settings module:
https://github.com/wizardsardine/liana/commit/5679255eb3df2d9565c6ab733a65595620b567ed

A second commit in v4 checked that settings.json is present
otherwise it does the migration from the gui configuration file
to the settings file:
https://github.com/wizardsardine/liana/commit/7381154dcb67a8f9bf130f86e75275560650d556

We remove the legacy field from the config field so the wallet module
does not depend on the config module anymore to load hardware wallet
information.