Merge launcher and first installer step
When the gui launches, it checks if there is
a wallet in the network datadirs with bitcoin mainnet in priority.
If it is the case, it tries to display information about
the wallet by reading the settings.json file.
If no wallet was found, then the gui displays the choice
to either create a wallet or add it.