gui: refresh cache more often while syncing
Refresh the cache more often while the wallet has a syncing
status of some kind in order to detect sooner that this has
finished.

This does not cover syncing scenarios that require the first poll
to be detected since we don't currently store the required info
(the last poll at startup). This should be covered in a follow-up.