Remove the link between gui config and lianad config
This commit is part of preparatory work to support
multiple wallets.

This commit introduces two breaking change:

The fields daemon_config_path and daemon_rpc_path are removed.
The GUI will deduce these values from their expected
location in the root data directory.

Because this link is removed, the gui flag --conf is not useful
anymore as it cannot then find the location of the root data directory.