refactor: move extension trait to http module
This trait will be useful in various other modules.

The `json_or_error` trait method has been removed as it can be
achieved by using Ok(...?) and trying to include it in the http
module with a generic error type can then lead to type
annotations being needed when using it due to several error
types implementing `From<reqwest::Error>`. I felt it was therefore
simpler to use the Ok(...?) approach.