Add global tick subscription
In order to reduce back pressure with multiple pane open,
only one subscription is providing a tick every 1 seconds.
It is up to each state of each panel to handle this tick.

For example the app state will compare the tick instant
to the last time it had updated the daemon cache to know
if it has to update it again.