fix: use Instant::now() in on_tick method
The tick parameter could be for an instant in the past, e.g. if
the GUI window is inactive for a while.

In the case of DaemonCache, storing last_tick would therefore
not prevent multiple updates from a burst of tick messages.

The on_tick() methods should rather use Instant::now() for
performing their logic.