use tick to trigger all fiat price requests
Given the global tick fires every second, we can use this to
handle all fiat price requests, including when the application
starts and when a user changes fiat settings.

It also allows to combine the logic that controls how often regular
updates are performed with that used to prevent multiple requests
being sent, which is done by using last_request in both cases.